[nrf noup] Aligned CI workflows for sdk-connectedhomeip
Add a manually triggered workflow for building Android
CHIPTool packages for arm and arm64 architectures,
CHIP Tool for Linux x64 and arm64, and OTA provider for
Linux.
The packages are uploaded both as github artifacts and
release packages (as the latter require a release tag).

The doxygen workflow pushes doxygen HTML files to another
branch which is undesired on the sdk-connectedhomeip fork
as it causes significant growth of the repository size.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>
Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>

Remove Android Chip Tool from release workflow

Android Chip Tool is no longer maintained so it must be
removed from the release workflow.

Remove dependency to openssl from CHIP tool build for linux

Currently the chip-tool that is released as a part of
sdk-connectedhomeip has dependency to openssl.
The issue is that Ubuntu 20.04 and Ubuntu 22.04 use different version
of OpenSSL and version built under 20.04 does not work under
22.04 (without some hacks to install the old openssl version)
and vice-verse.

- Removed the dependency to the dynamic library (OpenSSL) to build
Chip-Tool with the mbedTLS.

[nrf noup] Use "continue-on-error" for examples in CI

Make building nRF Connect examples optional in github
workflows in a sense that they do not fail the workflow
if they do not compile. The reason is that the CI is often
run for downstream patches, that are required to proceed
with changes in NCS or Zephyr and we do not want to force
developers to update examples in sdk-connectedhomeip at
that moment - they will be updated in the upstream later.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>

[nrf noup] Conditionally disable all unnecessary workflows

Currently, in sdk-connectedhomeip we don't need to run
upstream's workflows for all platforms because we use
sdk-nrf workflows to verify our PRs.

We decided to remove all of them except Lint Code Base and
Unit Tests from the 'pull request' and 'push' triggers and
leave the possibility to run them manually.

This commit cannot be pushed to the Matter upstream repository!

[nrf noup] Disable dependabot

It runs on the upstream repository and updates dependencies accordingly.

[nrf noup] Repaired unit tests

Last changes and upmerges cause instabilities in our workflows
actions. This commit repairs unit tests.

[nrf noup] [nrfconnect] Fix release tools workflow

[nrf noup] Aligned release tools to Matter 1.2

Added all changes to release tools github workflow needed to work
with Matter 1.2.

(cherry picked from commit 2b8d0377c964f70420b30a4ea56df28d4a1950eb)
Signed-off-by: Adrian Gielniewski <adrian.gielniewski@nordicsemi.no>