[nrfconnect] Introduce WiFi connection recover feature
The WiFi connection recover feature allows device re-scanning
and re-connecting to the known WiFi network after the device's reboot
and when the known SSID has not been found during the last scan.
The connection recover interval is doubled with
every occurrence to the defined maximum value and then its value
depends on the maximum value +- the defined random jitter.

After restoring the connection, the Connection Recovery
Interval is restored to the defined minimum value after elapsing
of the defined delay to avoid frequent reconnections to
poor link quality network.