[nrf noup] Increase Main thread priority to fix an app core crash
Currently, we have set 0 priority both to the main thread and
the nrf53 cpuapp IPC0 priority. When we perform a long operation
on the main thread we block IPC communication and get a crash on
app core when it tries to send a packet through ieee820154 radio.

To avoid that situation and the crash we can increase the main
thread priority to 1.