[nrfconnect] Introduced several improvements for the factory data management (#33195)
* [nrfconnect] Added FactoryDataBase to facilitate dynamic
assignments of FactoryDataProvider concrete types to
to the generic pointer.

Signed-off-by: Marcin Kajor <marcin.kajor@nordicsemi.no>

* [nrfconnect] Added non-PM based factory data support

Added support for platforms not built using PM. This commit
includes fixes only firmware implementation and does not include
cmake changes for factory data generation.

Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>

* [nrfconnect] Improved generation of the factory data

Allow to merge the factory data .hex file with the firmware's .hex
file when partition manager is not available.

* [nrfconnect] Disable External FD if QSPI NOR is not available

We need to have support for factory data when the partition manager
is enabled but there is no support for QSPI external flash.

---------

Signed-off-by: Marcin Kajor <marcin.kajor@nordicsemi.no>
Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>
Co-authored-by: Marcin Kajor <marcin.kajor@nordicsemi.no>
Co-authored-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>