Make generate.py a bit smarter about determining the ZCL definition file. (#33298)
If the ZCL definition file is not specified, it tries to look in the ZAP file
provided, but falls back to our in-tree zcl.json if the ZAP file has nothing
specified.

This change just uses the fallback if neither the ZCL definition file _nor_ the
ZAP file are specified, which allows simpler codegen command lines for
client-side things that don't have a specific ZAP file attached to them.