Conformance: add start of arithmetic conformance (#33424)
* Conformance: add start of arithmetic conformance

https://github.com/CHIP-Specifications/connectedhomeip-spec/pull/7808
represents the first arithmetic conformance operation. Adding
a new conformance class to match this. Currently this just returns
optional to match the current behaviour, but this will be changed
in an upcoming PR (see
https://github.com/project-chip/connectedhomeip/issues/33422)

However, thet change requires that we assess the conformance based
on the attribute values, which we do pass in right now. This
therefore requires a bit of an API change on the conformance classes
and I've opted to address this in a separate PR to keep the changes
compact.

* linter