Documentation: Building your first example (#33536)
* Documentation: Add MD version of webinar

Adds a slightly cleaned up version of the quick start guide
webinar section 1 as a quick start guide. The old quick start
is quite out of date, we should consider removing that and using
this as a replacement.

NOTE: Still need to hook in the indexes, and determin what to do
with the current quick start file (move this and replace?)

* add to index

* fix ref

* md files hate braces