[Python] Drop chip-device-ctrl (#33488)
* Drop chip-device-ctrl and ZCL* APIs

Drop the deprecated chip-device-ctrl and remove the ZCL* API from the
Python CHIP controller.

* Update docs to reflect chip-repl commands

Update the Python CHIP controller docs to reflect the CHIP REPL
instead of the now removed chip-device-ctrl tool.

* Remove unused imports

* Replace chip-device-ctrl with chip-repl

* Update/reword main QUICK_START and READMEs

* Fix wrong/buggy cross-reference

* Remove common chip-device-ctrl example

Remove the outdated chip-device-ctrl example and refer to the Python
controller REPL documentation.

* Add --ble-adapter support to CHIP REPL and update docs

Add support to select the Bluetooth adapter using the common
--ble-adapter command line argument.

Update the advanced docs for the Python Controller.

* Address review feedback

* Trim list of commands/add link to official API docs