TC-OPCREDS-3.2: Add (#32182)
* chore(TC_OPCREDS-3.2): skeleton base class

* chore(TC_OPCREDS_3.2): NOCResponse command

* chore(TC_OPCREDS_3.2): get currentFabric

* chore(TC_OPCREDS_3.2): print some status values

* TC-OPCREDS-3.2: Re-work test

This test was difficult to automate and included a factory reset.
This automation start is used to generate the test plans.

The intent of this test is to confirm that the returned fabric
index matches the fabric index returned from the AddNOC command,
and that the read of fabric-filtered attributes matches to the
calling fabric. The rest of the test is covered by OPCREDS-3.1
so was removed to simplify.

NOTE: test not yet implemented, uploaded for test plan generation
      purposes only.

- Properly use Test Step and Expected outcome columns
- Remove checks for NOCs matching - this is checked in OPCREDS-3.1
- Remove factory reset and commission two new fabrics
- Use commissioning over CASE so we can more easily return the
  fabric index

* chore(TC_OPCREDS_3.2): included new steps (1-6)

Just need to add validation for CR3 and remove all fabrics. Also
it need clean up (remove prints, unused variables, etc)

* chore(TC_OPCREDS_3.2): all steps are implemtented

* chore(TC_OPCREDS_3.2): restyled by isort

* chore(TC_OPCREDS_3_2): fix error from python linter (code-lints)

* chore(TC_OPCREDS_3.2): import restyled by isort

* chore(TC_OPCREDS_3.2): fixed the vendorID for prevent fail

* chore(TC_OPCREDS_3.2): add test to CI Workflow

* chore(TC_OPCREDS_3.2): off auto-format and change the return values label

* chore(TC_OPCREDS_3.2): remove if-else validation instead used assert_equal

* chore(TC_OPCREDS_3.2): spacing adjustment for better readability

* chore(TC_OPCREDS_3.2): identation adjustment for better readability

* chore(TC_OPCREDS_3.2): restyled patch identation adjustment

* chore(TC_OPCREDS_3.2): reverted addition of test

---------

Co-authored-by: cecille <cecille@google.com>