Close TCP connection when received message size is too large. (#33768)
When the framing length value of a received message is larger
than what the local node can process, abort the connection
with the peer.

Sending a StatusResponse message back to the peer as a
notification may not be feasible in all circumstances for
reasons, such as:
1) It would require a cross-layered feedback up to the Exchange
   layer to generate such a message in response to a failure at the
   transport layer.
2) A Status Response is sent in response to a message on an
   ExchangeContext and that may not be the case in scnearios
   where this message is the first unsolicited message.

The receiver could drain out the bits from the offending message
and move on to the next message in the stream but that may not
guarantee correct behavior and would consume resources unnecessarily.

Given that the peer was already aware of the max length this node was
willing to receive during its TCP advertisement, it seems prudent
to fail fast and close the connection.

Fixes #33307.