Add SupportedCluster list to ContentAppPlatform (#33827)
* Add SupportedCluster list to ContentAppPlatform

[Problem]
The ContentAppPlatform does not know which clusters each ContentApp supports.

Currently the ContentApp is queried to get the passcode as long as the
AccountLoginDelegate is present whether or not the installed ContentApp
has declared support for AccountLogin cluster in its manifest.

[Solution]
Extend the native ContentApp with a SupportedCluster list. The list is
initialized from the Android AppPlatformService whenever an installed ContentApp
is discovered and added to the native AppPlatform.

This list is used to check if AccountLogin cluster is supported before
querying the ContentApp for the passcode.

[Test]
The feature is tested end-to-end using a native Linux casting-app and
and an Android platform-app and content-app. The content-app
static_matter_clusters raw asset was manipulated to verify that the
clusters are parsed correctly and that the passcode is only retrievable
when the AccountLogin cluster is declared.

* Restyled by whitespace

* Restyled by google-java-format

* Restyled by clang-format

* ContentApp default constructor

* Fix code-lints: check_includes

* mapSupportedClusters type fix

* Rename feats to features

* AccountLogin check in HastargetContentApp

* Rename SupportedCluster members to disambiguate

* AppPlatformService addContentAppAtEndpoint argument fix

* Default SupportedCluster init

* Restyled by clang-format

* Add AccountLogin cluster to static_matter_clusters

* Remove unnecessary namespace specifiers

---------

Co-authored-by: Restyled.io <commits@restyled.io>