ReadClient: Truncate data version list during encoding if necessary (#34111)
* ReadClient: Truncate data version list during encoding if necessary

The existing code made the assumption that if a list of versions was able to
fit into the request packet when generating the first subscribe request, then
any resubscribe containing data versions for the same clusters would also fit.
However the data version numbers themselves can be updated when we receive
reports, and this can cause the list to no longer fit the request packet,
leaving us in a state where every resubscribe attempt would fail.

Note that this change means even an initial subscribe request with a data
version list that is too long will no longer fail; ReadClient will simply
truncate the list as needed in all cases.

* Apply comment suggestions from code review

Co-authored-by: Boris Zbarsky <bzbarsky@apple.com>

* Treat CHIP_ERROR_BUFFER_TOO_SMALL the same

* Switch data_model tests to pw_unit_test

* Add WillSendMessage to loopback delegate and make source addresses more plausible

* Add test for ReadClient data version truncation

* Make the linter happy

---------

Co-authored-by: Boris Zbarsky <bzbarsky@apple.com>